Don't invert Y, actually.
This is correct because this edge of the bend goes right to left _and_
sets end alignment. These cancel each other out, ensuring that the short
step that embattled starts with (which is the beginning of the pattern)
is always on the left side, and always mirroring the top side. Draw
it out if you don't believe me.