Fix footnotes in the bread recipe.
The problem was that the parser was treating the content of the block
as the _whole_ document, because of the way that the Jekyll Kramdown
parser wrapper is implemented:

https://github.com/jekyll/jekyll/blob/a6b861cd9509002b8da7b4d3168d79a84a86c6ba/lib/jekyll/converters/markdown/kramdown_parser.rb#L109

Note that it makes a new thing and then HTMLifies it. I thought it was
more incremental.

The fix was to make the tag non-block, and instead just inject the
headers, which really makes more sense anyway. (The only loss is the
ability to subtly highlight the background of the step... I guess I
could manually wrap them in a `<section>` if I really wanted that.)