After much experimentation with DRYing up footnotes, this is the best it's going to get.
While Kramdown permits the `{:footnotes}` reference to allow moving
footnotes around in the page, pretty much the only way this can be used
outside of the post content itself is with a `{% capture %}`, and that's
pretty verbose:

```
{% capture footnotes %}
* dummy list
{:footnotes}
{% endcapture %}
```

Yes, the list is necessary. Only 1 of those newlines in there is
_strictly_ necessary, but it's even worse boilerplate if you try to
collapse it.

Anyway, back to the main thread: captures can be used in includes, but
they cannot be used in layouts, for some reason. So while a straight
`{:footnotes}` reference can be moved around within a post, and a
capture + reference combination can be used to exfiltrate the footnotes
to an included partial, you still need that include and can't hoist the
footnotes up to the layout to have all the various page bits ordered in
one place. Lame.

Another tack I took was with the `jekyll-contentblocks` plugin which
acts kind of like Rails' `content_for`. It doesn't work, because the
content inside the block tag is rendered in a different context, so
Kramdown loses the link between footnote references and definition, and
complains that the footnote goes nowhere. Maybe the plugin can use some
other API to render in the same context, but I doubt it. Perhaps the
implementation of the `capture` tag has a hint?