Set pattern sizing for some SVGs.
This is done by rendering one copy of the charge with something like

> argent a <charge> or

and then using `getBoundingClientRect` to get the calculated size,
making sure to first remove any the scaling transform introduced by
the renderer for having a single charge, and then dividing the sizing
by 3 (which is the current multiplication factor from logical SVG size
to rendered SVG size in the DOM).

If you ask for the bounding rect when examining the SVG in isolation,
Firefox will just report either the stated width/height, if present, or
the current viewport size otherwise. Neither of these are useful.